стилізував ховери в галереї і пагінації
