+++ src/partials/cocktails.html
@@ -1,46 +1,16 @@
 <section class="cocktails">
   <div class="container">
     <h2 class="cocktails__title">Cocktails</h2>
-    <ul class="gallery">
-      <!-- <li class="gallery__item">
-        <img
-            class="gallery__img"
-            src="./images/Vampiro.jpg"
-            alt="Negroni"
-            width="280px"
-            height="280px"
-          />
-          <h3 class="gallery__subtitle">Negroni</h3>
-        <div class="gallery__btns">
-          <button class="btn btn--orange" type="button">Learn more</button>
-          <button class="btn btn--white" type="button">
-            Add to &nbsp
-            <span class="btn__icon-wrap">
-              <svg class="btn__icon" width="15" height="15">
-                <use href="./images/sprite.svg#heart"></use>
-              </svg>
-            </span>
-          </button>
-          <button class="btn btn--white visually-hidden" type="button">
-            Remove &nbsp
-            <span class="btn__icon-wrap"> </span>
-          </button>
-        </div>
-      </li> -->
-    </ul>
+    <ul class="gallery"></ul>
     <div class="pagination visually-hidden">
       <span class="pagination__previous">
-        <svg class="pagination__icon pagination--disabled" width="8" height="13">
+        <svg class="pagination__icon pagination--disabled" width="8" height="14">
           <use href="./images/sprite.svg#chevronleft"></use>
         </svg>
       </span>
-      <ul class="pagination__list">
-        <!-- <li class="pagination__item pagination--active">1</li>
-        <li class="pagination__item">2</li>
-        <li class="pagination__item">3</li> -->
-      </ul>
+      <ul class="pagination__list"></ul>
       <span class="pagination__next">
-        <svg class="pagination__icon" width="8" height="13">
+        <svg class="pagination__icon" width="8" height="14">
           <use href="./images/sprite.svg#chevronright"></use>
         </svg>
       </span>
