+++ src/sass/components/_cocktails.scss
@@ -38,6 +38,22 @@
     display: flex;
     align-items: center;
     justify-content: center;
+    width: 24px;
+    height: 24px;
+    border-radius: 50%;
+    cursor: pointer;
+    transition: background-color 250ms var(--timing-function);
+  }
+  &__previous:hover,
+  &__next:hover {
+    background-color: var(--auxiliary-orange);
+  }
+  .pagination__icon {
+    transition: fill 250ms var(--timing-function);
+  }
+  &__previous:hover .pagination__icon,
+  &__next:hover .pagination__icon {
+    fill: var(--white);
   }
   &__list {
     display: flex;
@@ -55,12 +71,18 @@
     font-family: 'Montserrat', sans-serif;
     border: 1px solid var(--auxiliary-orange);
     border-radius: 50%;
+    transition: color 250ms var(--timing-function);
+    transition: background-color 250ms var(--timing-function);
     cursor: pointer;
     @include media-min($tablet) {
       width: 40px;
       height: 40px;
       font-size: 16px;
     }
+  }
+  &__item:hover {
+    color: var(--white);
+    background-color: var(--grapefruit);
   }
   &--active {
     color: var(--white);
